fix: 잘못된 파일 경로 수정
